trainer/pyproject.toml
[project]
name = "trainer"
version = "0.1.0"
description = "Vertex AI | Custom Training | AutoGluon"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "autogluon>=1.3.0",
    "click>=8.1.8",
    "gcsfs>=2025.3.2",
    "google-cloud-bigquery>=3.31.0",
    "google-cloud-storage>=3.1.0",
    "msgspec>=0.19.0",
    "pandas>=2.2.3",
    "scikit-learn>=1.6.1",
]

[dependency-groups]
dev = ["bandit>=1.8.0", "mypy>=1.13.0", "ruff>=0.8.2"]

[tool.ruff]
line-length = 80

[tool.mypy]
ignore_missing_imports = true
disable_error_code = ["import-untyped", "attr-defined"]

[tool.bandit]
exclude_dirs = ["tests"]
